Разбивает код на отдельные модули
