+++ js/util.js
@@ -1,0 +1,37 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+function createIdGenerator() {
+  let lastGeneratedId = 1;
+
+  return function() {
+    return lastGeneratedId++;
+  };
+}
+
+function createRandomUniqueIdGenerator(min, max) {
+  const previousValues = [];
+
+  return function() {
+    let currentValue = getRandomInteger(min, max);
+
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+export {getRandomInteger, getRandomArrayElement, createIdGenerator, createRandomUniqueIdGenerator};
